Restructure Dockerfiles
In other pipelines, it works reasonably well to have one container that's just
software (ours and third-party packages), but this has proven to be
sub-optimal for this pipeline. Each small tweak to our code produces another
5.6GB of data that can't take advantage of the Docker build cache.

Fix this by flipping the order: make a "base" container that's just
rocker/r-ver:3.6.3 plus our data dependencies, then build on that to add
software packages and our code.